Moved cmd and pkg source codde into src folder. Updated fsm pkg import in main.go
